// lib/api/model/attendanceMethod.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type AttendanceMethod = typeof AttendanceMethod[keyof typeof AttendanceMethod];


export const AttendanceMethod = {
  None: 'None',
  Manual: 'Manual',
  Nfc: 'Nfc',
  Qr: 'Qr',
  QrWifi: 'QrWifi',
} as const;