// lib/api/model/attendanceStatus.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  Present: 'Present',
  Late: 'Late',
  Absent: 'Absent',
  Excused: 'Excused',
} as const;